requirements.txt
discord.py[voice]>=2.3.2
aiohttp>=3.8.0
aiofiles>=23.0.0
python-dotenv>=1.0.0
pytest>=7.0.0
pytest-asyncio>=0.21.0